server/model/requirements.txt
pandas
numpy
click
matplotlib
keras
tensorflow-gpu
tensorboard
pydot
graphviz
h5py
